migrate dmn-example3 to maven profile
