+++ .github/workflows/mvn-pam-quick-examples.yaml
@@ -60,8 +60,8 @@ jobs:
           mvn compile --no-transfer-progress
           popd
 
-          pushd pam-quick-examples/dmn-example3
-          bash build.sh
+          pushd pam-quick-examples/dmn-example3/dmn-example3-parent
+          mvn compile --no-transfer-progress
           popd
 
           pushd pam-quick-examples/dmn-example4
