+++ pam-quick-examples/dmn-example1/dmn-example1-parent/dmn-example1-kie-api-remote-impl/src/main/java/com/redhat/cop/pam/example1/kie/api/impl/RulesApiImpl.java
@@ -35,7 +35,7 @@ public class RulesApiImpl implements RulesApi {
 
     private static final String PASSWORD = System.getProperty("com.redhat.cop.pam.kieserver_password");
 
-    private static final String CONTAINER_ID = System.getProperty("remote.kieserver_containerid");
+    private static final String CONTAINER_ID = System.getProperty("com.redhat.cop.pam.kieserver_containerid");
 
     private static final String NAMESPACE = "https://pam.cop.redhat.com/dmn-example-1/dmn";
