--- pam-quick-examples/dmn-example3/build.sh
@@ -1,22 +1,0 @@
-#!/usr/bin/env bash
-
-EXAMPLE_PATH="$(cd "$(dirname "$0")" && pwd)"
-EXAMPLE_ID="$(basename "$EXAMPLE_PATH")"
-
-declare -a example_projects=("parent" "kjar" "quarkus-embedded" "quarkus-remote" "springboot-embedded" "springboot-remote")
-
-function maven_clean_install_skip_tests(){
-    POM_ABSOLUTE_PATH="$EXAMPLE_PATH/$EXAMPLE_ID-$1/pom.xml"
-    echo "$POM_ABSOLUTE_PATH"
-    mvn clean install -f "$POM_ABSOLUTE_PATH" -DskipTests --no-transfer-progress
-}
-
-function build_projects(){
-    for i in "${example_projects[@]}"
-        do
-        :
-        maven_clean_install_skip_tests "$i"
-    done
-}
-
-build_projects
