+++ pam-quick-examples/dmn-example3/dmn-example3-parent/dmn-example3-kie-api-remote-impl/src/main/java/com/redhat/cop/pam/example3/kie/api/impl/RulesApiImpl.java
@@ -7,6 +7,8 @@ import org.kie.dmn.api.core.DMNContext;
 import org.kie.dmn.api.core.DMNDecisionResult;
 import org.kie.dmn.api.core.DMNMessage;
 import org.kie.dmn.api.core.DMNResult;
+import org.kie.server.api.model.KieContainerResource;
+import org.kie.server.api.model.ReleaseId;
 import org.kie.server.api.model.ServiceResponse;
 import org.kie.server.api.model.instance.ProcessInstance;
 import org.kie.server.client.*;
@@ -14,6 +16,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.annotation.PostConstruct;
+import javax.annotation.PreDestroy;
 import javax.enterprise.context.ApplicationScoped;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -26,13 +29,13 @@ public class RulesApiImpl implements RulesApi {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(RulesApiImpl.class);
 
-    private static final String URL = System.getProperty("com.redhat.cop.pam.kieserver_url", "http://localhost:8080/kie-server/services/rest/server");
+    private static final String URL = System.getProperty("com.redhat.cop.pam.kieserver_url");
 
-    private static final String USER = System.getProperty("com.redhat.cop.pam.kieserver_user", "kieServerUser");
+    private static final String USER = System.getProperty("com.redhat.cop.pam.kieserver_user");
 
-    private static final String PASSWORD = System.getProperty("com.redhat.cop.pam.kieserver_password", "kieServerUser1234;");
+    private static final String PASSWORD = System.getProperty("com.redhat.cop.pam.kieserver_password");
 
-    private static final String CONTAINER_ID = "com.redhat.cop.pam:dmn-example3-kjar:1.0";
+    private static final String CONTAINER_ID = System.getProperty("com.redhat.cop.pam.kieserver_containerid");
 
     private KieServicesClient kieServicesClient;
 
@@ -44,6 +47,14 @@ public class RulesApiImpl implements RulesApi {
         config.addExtraClasses(extraClasses);
         config.setTimeout(100000l);
         kieServicesClient = KieServicesFactory.newKieServicesClient(config);
+        final String[] gav = CONTAINER_ID.split(":");
+        final KieContainerResource containerResource = new KieContainerResource(CONTAINER_ID, new ReleaseId(gav[0], gav[1], gav[2]));
+        kieServicesClient.createContainer(CONTAINER_ID, containerResource);
+    }
+
+    @PreDestroy
+    public void preDestroy(){
+        kieServicesClient.disposeContainer(CONTAINER_ID);
     }
 
     @Override
