+++ pam-quick-examples/dmn-example3/dmn-example3-parent/dmn-example3-kie-api-remote-impl/src/main/java/com/redhat/cop/pam/example3/kie/api/utils/KieContainerUtils.java
@@ -1,0 +1,4 @@
+package com.redhat.cop.pam.example3.kie.api.utils;
+
+public class KieContainerUtils {
+}
