+++ pam-quick-examples/dmn-example3/dmn-example3-parent/dmn-example3-kjar/project.imports
