+++ pam-quick-examples/dmn-example3/dmn-example3-parent/dmn-example3-kjar/src/main/resources/com/redhat/cop/pam/example3/kjar/dmn/customer-can-open-account.dmn
