+++ pam-quick-examples/dmn-example3/dmn-example3-parent/dmn-example3-quarkus/pom.xml
@@ -1,0 +1,93 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>com.redhat.cop.pam</groupId>
+        <artifactId>dmn-example3-parent</artifactId>
+        <version>1.0</version>
+    </parent>
+
+    <artifactId>dmn-example3-quarkus</artifactId>
+    <name>${project.artifactId}-${project.version}</name>
+    <packaging>jar</packaging>
+
+    <dependencies>
+        <dependency>
+            <groupId>com.redhat.cop.pam</groupId>
+            <artifactId>dmn-example3-services-api</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.redhat.cop.pam</groupId>
+            <artifactId>dmn-example3-services-api-impl</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.redhat.cop.pam</groupId>
+            <artifactId>dmn-example3-kie-api</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>io.quarkus</groupId>
+            <artifactId>quarkus-resteasy</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>io.quarkus</groupId>
+            <artifactId>quarkus-jsonb</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>io.quarkus</groupId>
+            <artifactId>quarkus-resteasy-jsonb</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>io.quarkus</groupId>
+            <artifactId>quarkus-smallrye-openapi</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>io.quarkus</groupId>
+            <artifactId>quarkus-junit5</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>io.rest-assured</groupId>
+            <artifactId>rest-assured</artifactId>
+            <scope>test</scope>
+        </dependency>
+    </dependencies>
+
+    <profiles>
+        <profile>
+            <id>remote</id>
+            <dependencies>
+                <dependency>
+                    <groupId>com.redhat.cop.pam</groupId>
+                    <artifactId>dmn-example3-kie-api-remote-impl</artifactId>
+                </dependency>
+            </dependencies>
+            <build>
+                <plugins>
+                    <plugin>
+                        <groupId>org.apache.maven.plugins</groupId>
+                        <artifactId>maven-surefire-plugin</artifactId>
+                        <configuration>
+                            <systemPropertyVariables>
+                                <com.redhat.cop.pam.kieserver_url>${remote.kieserver_url}</com.redhat.cop.pam.kieserver_url>
+                                <com.redhat.cop.pam.kieserver_user>${remote.kieserver_user}</com.redhat.cop.pam.kieserver_user>
+                                <com.redhat.cop.pam.kieserver_password>${remote.kieserver_password}</com.redhat.cop.pam.kieserver_password>
+                                <com.redhat.cop.pam.kieserver_containerid>${remote.kieserver_containerid}</com.redhat.cop.pam.kieserver_containerid>
+                            </systemPropertyVariables>
+                        </configuration>
+                    </plugin>
+                </plugins>
+            </build>
+        </profile>
+        <profile>
+            <id>embedded</id>
+            <dependencies>
+                <dependency>
+                    <groupId>com.redhat.cop.pam</groupId>
+                    <artifactId>dmn-example3-kie-api-embedded-impl</artifactId>
+                </dependency>
+            </dependencies>
+        </profile>
+    </profiles>
+</project>
