+++ pam-quick-examples/dmn-example3/dmn-example3-parent/dmn-example3-quarkus/src/test/java/com/redhat/cop/pam/example3/quarkus/ServiceTest.java
