+++ pam-quick-examples/dmn-example3/dmn-example3-parent/dmn-example3-springboot/pom.xml
@@ -1,0 +1,112 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>com.redhat.cop.pam</groupId>
+        <artifactId>dmn-example3-parent</artifactId>
+        <version>1.0</version>
+    </parent>
+
+    <artifactId>dmn-example3-springboot</artifactId>
+    <name>${project.artifactId}-${project.version}</name>
+    <packaging>jar</packaging>
+
+    <dependencies>
+        <dependency>
+            <groupId>com.redhat.cop.pam</groupId>
+            <artifactId>dmn-example3-services-api</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.redhat.cop.pam</groupId>
+            <artifactId>dmn-example3-services-api-impl</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.redhat.cop.pam</groupId>
+            <artifactId>dmn-example3-kie-api</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-web</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.cxf</groupId>
+            <artifactId>cxf-spring-boot-starter-jaxrs</artifactId>
+            <version>3.3.4.redhat-00001</version>
+        </dependency>
+        <dependency>
+            <groupId>org.codehaus.jackson</groupId>
+            <artifactId>jackson-jaxrs</artifactId>
+            <version>1.9.13.redhat-4</version>
+        </dependency>
+        <dependency>
+            <groupId>org.codehaus.jackson</groupId>
+            <artifactId>jackson-xc</artifactId>
+            <version>1.9.13.redhat-4</version>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-test</artifactId>
+            <scope>test</scope>
+        </dependency>
+    </dependencies>
+
+    <profiles>
+        <profile>
+            <id>remote</id>
+            <dependencies>
+                <dependency>
+                    <groupId>com.redhat.cop.pam</groupId>
+                    <artifactId>dmn-example3-kie-api-remote-impl</artifactId>
+                </dependency>
+            </dependencies>
+            <build>
+                <plugins>
+                    <plugin>
+                        <groupId>org.apache.maven.plugins</groupId>
+                        <artifactId>maven-surefire-plugin</artifactId>
+                        <configuration>
+                            <systemPropertyVariables>
+                                <com.redhat.cop.pam.kieserver_url>${remote.kieserver_url}</com.redhat.cop.pam.kieserver_url>
+                                <com.redhat.cop.pam.kieserver_user>${remote.kieserver_user}</com.redhat.cop.pam.kieserver_user>
+                                <com.redhat.cop.pam.kieserver_password>${remote.kieserver_password}</com.redhat.cop.pam.kieserver_password>
+                                <com.redhat.cop.pam.kieserver_containerid>${remote.kieserver_containerid}</com.redhat.cop.pam.kieserver_containerid>
+                            </systemPropertyVariables>
+                        </configuration>
+                    </plugin>
+                    <plugin>
+                        <groupId>org.springframework.boot</groupId>
+                        <artifactId>spring-boot-maven-plugin</artifactId>
+                        <configuration>
+                            <systemPropertyVariables>
+                                <com.redhat.cop.pam.kieserver_url>${remote.kieserver_url}</com.redhat.cop.pam.kieserver_url>
+                                <com.redhat.cop.pam.kieserver_user>${remote.kieserver_user}</com.redhat.cop.pam.kieserver_user>
+                                <com.redhat.cop.pam.kieserver_password>${remote.kieserver_password}</com.redhat.cop.pam.kieserver_password>
+                                <com.redhat.cop.pam.kieserver_containerid>${remote.kieserver_containerid}</com.redhat.cop.pam.kieserver_containerid>
+                            </systemPropertyVariables>
+                        </configuration>
+                    </plugin>
+                </plugins>
+            </build>
+        </profile>
+        <profile>
+            <id>embedded</id>
+            <dependencies>
+                <dependency>
+                    <groupId>com.redhat.cop.pam</groupId>
+                    <artifactId>dmn-example3-kie-api-embedded-impl</artifactId>
+                </dependency>
+            </dependencies>
+            <build>
+                <plugins>
+                    <plugin>
+                        <groupId>org.springframework.boot</groupId>
+                        <artifactId>spring-boot-maven-plugin</artifactId>
+                    </plugin>
+                </plugins>
+            </build>
+        </profile>
+    </profiles>
+</project>
