+++ pam-quick-examples/dmn-example3/dmn-example3-parent/dmn-example3-springboot/src/main/resources/application.properties
