+++ pam-quick-examples/dmn-example3/dmn-example3-parent/dmn-example3-springboot/src/test/java/com/redhat/cop/pam/example3/springboot/ServiceTest.java
