+++ pam-quick-examples/dmn-example3/dmn-example3-parent/pom.xml
@@ -21,11 +21,10 @@
     <modules>
         <module>dmn-example3-model</module>
         <module>dmn-example3-kie-api</module>
-        <module>dmn-example3-kie-api-remote-impl</module>
-        <module>dmn-example3-kie-api-embedded-impl</module>
         <module>dmn-example3-services-api</module>
         <module>dmn-example3-services-api-impl</module>
         <module>dmn-example3-functions</module>
+        <module>dmn-example3-kjar</module>
     </modules>
 
     <dependencyManagement>
@@ -37,17 +36,12 @@
             </dependency>
             <dependency>
                 <groupId>com.redhat.cop.pam</groupId>
+                <artifactId>dmn-example3-kjar</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>com.redhat.cop.pam</groupId>
                 <artifactId>dmn-example3-kie-api</artifactId>
-                <version>${project.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>com.redhat.cop.pam</groupId>
-                <artifactId>dmn-example3-kie-api-embedded-impl</artifactId>
-                <version>${project.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>com.redhat.cop.pam</groupId>
-                <artifactId>dmn-example3-kie-api-remote-impl</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
@@ -74,5 +68,128 @@
             </dependency>
         </dependencies>
     </dependencyManagement>
+
+    <profiles>
+        <profile>
+            <id>quarkus</id>
+            <properties>
+                <quarkus.bom.version>1.5.2.Final</quarkus.bom.version>
+                <quarkus-plugin.version>1.5.2.Final</quarkus-plugin.version>
+                <surefire-plugin.version>2.22.1</surefire-plugin.version>
+            </properties>
+            <modules>
+                <module>dmn-example3-quarkus</module>
+            </modules>
+            <dependencyManagement>
+                <dependencies>
+                    <dependency>
+                        <groupId>io.quarkus</groupId>
+                        <artifactId>quarkus-bom</artifactId>
+                        <version>${quarkus.bom.version}</version>
+                        <type>pom</type>
+                        <scope>import</scope>
+                    </dependency>
+                    <dependency>
+                        <groupId>com.redhat.cop.pam</groupId>
+                        <artifactId>dmn-example3-quarkus</artifactId>
+                        <version>${project.version}</version>
+                    </dependency>
+                </dependencies>
+            </dependencyManagement>
+            <build>
+                <plugins>
+                    <plugin>
+                        <groupId>io.quarkus</groupId>
+                        <artifactId>quarkus-maven-plugin</artifactId>
+                        <version>${quarkus-plugin.version}</version>
+                        <executions>
+                            <execution>
+                                <goals>
+                                    <goal>build</goal>
+                                </goals>
+                            </execution>
+                        </executions>
+                    </plugin>
+                    <plugin>
+                        <groupId>org.apache.maven.plugins</groupId>
+                        <artifactId>maven-surefire-plugin</artifactId>
+                        <version>${surefire-plugin.version}</version>
+                        <configuration>
+                            <systemProperties>
+                                <java.util.logging.manager>org.jboss.logmanager.LogManager</java.util.logging.manager>
+                            </systemProperties>
+                        </configuration>
+                    </plugin>
+                    <plugin>
+                        <groupId>io.quarkus</groupId>
+                        <artifactId>quarkus-bootstrap-maven-plugin</artifactId>
+                        <version>${quarkus-plugin.version}</version>
+                    </plugin>
+                </plugins>
+            </build>
+        </profile>
+        <profile>
+            <id>springboot</id>
+            <properties>
+                <swagger.version>2.9.2</swagger.version>
+                <surefire-plugin.version>2.22.1</surefire-plugin.version>
+            </properties>
+            <modules>
+                <module>dmn-example3-springboot</module>
+            </modules>
+            <dependencyManagement>
+                <dependencies>
+                    <dependency>
+                        <groupId>org.springframework.boot</groupId>
+                        <artifactId>spring-boot-starter-parent</artifactId>
+                        <version>2.2.2.RELEASE</version>
+                        <type>pom</type>
+                        <scope>import</scope>
+                    </dependency>
+                    <dependency>
+                        <groupId>com.redhat.cop.pam</groupId>
+                        <artifactId>dmn-example3-springboot</artifactId>
+                        <version>${project.version}</version>
+                    </dependency>
+                </dependencies>
+            </dependencyManagement>
+        </profile>
+        <profile>
+            <id>embedded</id>
+            <modules>
+                <module>dmn-example3-kie-api-embedded-impl</module>
+            </modules>
+            <dependencyManagement>
+                <dependencies>
+                    <dependency>
+                        <groupId>com.redhat.cop.pam</groupId>
+                        <artifactId>dmn-example3-kie-api-embedded-impl</artifactId>
+                        <version>${project.version}</version>
+                    </dependency>
+                </dependencies>
+            </dependencyManagement>
+        </profile>
+        <profile>
+            <id>remote</id>
+            <properties>
+                <remote.kieserver_url>http://localhost:8080/kie-server/services/rest/server</remote.kieserver_url>
+                <remote.kieserver_user>kieServerUser</remote.kieserver_user>
+                <remote.kieserver_password>kieServerUser1234;</remote.kieserver_password>
+                <remote.kieserver_containerid>com.redhat.cop.pam:dmn-example3-kjar:1.0</remote.kieserver_containerid>
+            </properties>
+            <modules>
+                <module>dmn-example3-kie-api-remote-impl</module>
+            </modules>
+            <dependencyManagement>
+                <dependencies>
+                    <dependency>
+                        <groupId>com.redhat.cop.pam</groupId>
+                        <artifactId>dmn-example3-kie-api-remote-impl</artifactId>
+                        <version>${project.version}</version>
+                    </dependency>
+                </dependencies>
+            </dependencyManagement>
+        </profile>
+    </profiles>
 
 </project>
