--- pam-quick-examples/dmn-example3/dmn-example3-quarkus-embedded/src/main/java/com/redhat/cop/pam/example3/Application.java
@@ -1,5 +1,0 @@
-package com.redhat.cop.pam.example3;
-
-public class Application {
-
-}
