--- pam-quick-examples/dmn-example3/dmn-example3-quarkus-remote/src/main/resources/application.properties
@@ -1,10 +1,0 @@
-# LOGGER
-quarkus.log.level=INFO
-
-# SMALLERYE-OPENAPI
-quarkus.smallrye-openapi.path=/swagger
-
-# HTTP
-quarkus.http.root-path=/dmn-example-2
-quarkus.http.port=8280
-quarkus.http.test-port=8280
