--- pam-quick-examples/dmn-example3/dmn-example3-springboot-embedded/pom.xml
@@ -1,92 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <groupId>com.redhat.cop.pam</groupId>
-    <artifactId>dmn-example3-springboot-embedded</artifactId>
-    <version>1.0</version>
-
-    <name>${project.artifactId}-${project.version}</name>
-    <packaging>jar</packaging>
-
-    <properties>
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <maven.compiler.target>1.8</maven.compiler.target>
-        <maven.compiler.release>1.8</maven.compiler.release>
-        <swagger.version>2.9.2</swagger.version>
-        <example3.parent.version>1.0</example3.parent.version>
-    </properties>
-
-    <dependencyManagement>
-        <dependencies>
-            <dependency>
-                <groupId>com.redhat.cop.pam</groupId>
-                <artifactId>dmn-example3-parent</artifactId>
-                <version>${example3.parent.version}</version>
-                <type>pom</type>
-                <scope>import</scope>
-            </dependency>
-            <dependency>
-                <groupId>org.springframework.boot</groupId>
-                <artifactId>spring-boot-starter-parent</artifactId>
-                <version>2.2.2.RELEASE</version>
-                <type>pom</type>
-                <scope>import</scope>
-            </dependency>
-        </dependencies>
-    </dependencyManagement>
-
-    <dependencies>
-        <dependency>
-            <groupId>com.redhat.cop.pam</groupId>
-            <artifactId>dmn-example3-services-api</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.redhat.cop.pam</groupId>
-            <artifactId>dmn-example3-services-api-impl</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.redhat.cop.pam</groupId>
-            <artifactId>dmn-example3-kie-api</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.redhat.cop.pam</groupId>
-            <artifactId>dmn-example3-kie-api-embedded-impl</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-web</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.cxf</groupId>
-            <artifactId>cxf-spring-boot-starter-jaxrs</artifactId>
-            <version>3.3.4.redhat-00001</version>
-        </dependency>
-        <dependency>
-            <groupId>org.codehaus.jackson</groupId>
-            <artifactId>jackson-jaxrs</artifactId>
-            <version>1.9.13.redhat-4</version>
-        </dependency>
-        <dependency>
-            <groupId>org.codehaus.jackson</groupId>
-            <artifactId>jackson-xc</artifactId>
-            <version>1.9.13.redhat-4</version>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-test</artifactId>
-            <scope>test</scope>
-        </dependency>
-    </dependencies>
-
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.springframework.boot</groupId>
-                <artifactId>spring-boot-maven-plugin</artifactId>
-            </plugin>
-        </plugins>
-    </build>
-</project>
