--- pam-quick-examples/dmn-example3/dmn-example3-springboot-embedded/src/main/java/com/redhat/cop/pam/example3/springboot/Application.java
@@ -1,27 +1,0 @@
-package com.redhat.cop.pam.example3.springboot;
-
-import com.redhat.cop.pam.example3.kie.api.RulesApi;
-import com.redhat.cop.pam.example3.kie.api.impl.RulesApiImpl;
-import com.redhat.cop.pam.example3.kie.api.utils.KieContainerUtils;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-
-@SpringBootApplication
-public class Application {
-
-    public static void main(String[] args) {
-        SpringApplication.run(Application.class, args);
-    }
-
-    @Bean
-    public KieContainerUtils initKieContainerUtils() {
-        return new KieContainerUtils();
-    }
-
-    @Bean
-    public RulesApi initRulesApi() {
-        return new RulesApiImpl();
-    }
-
-}
