--- pam-quick-examples/dmn-example3/dmn-example3-springboot-remote/src/main/resources/application.properties
@@ -1,6 +1,0 @@
-server.port=8280
-
-cxf.path=/dmn-example-3
-cxf.jaxrs.classes-scan=true
-cxf.jaxrs.classes-scan-packages=com.redhat.cop.pam.example3.services.api,org.codehaus.jackson.jaxrs
-
