--- pam-quick-examples/dmn-example3/requests/com.redhat.cop.pam_dmn-example3-kjar_1.0.json
@@ -1,33 +1,0 @@
-{
-    "config-items": [
-      {
-        "itemName": "RuntimeStrategy",
-        "itemValue": "SINGLETON",
-        "itemType": "java.lang.String"
-      },
-      {
-        "itemName": "MergeMode",
-        "itemValue": "MERGE_COLLECTIONS",
-        "itemType": "java.lang.String"
-      },
-      {
-        "itemName": "KBase",
-        "itemValue": "",
-        "itemType": "java.lang.String"
-      },
-      {
-        "itemName": "KSession",
-        "itemValue": "",
-        "itemType": "java.lang.String"
-      }
-    ],
-    "release-id": {
-      "group-id": "com.redhat.cop.pam",
-      "artifact-id": "dmn-example3-kjar",
-      "version": "1.0"
-    },
-    "scanner": {
-      "poll-interval": "5000",
-      "status": "STARTED"
-    }
-  }
